Fix to remove dependency on portable_minimal_ada_rts crate
HiRTOS does not build with the published portable_minimal_ada_rts
crate. Some symbols were removed from the published portable_minimal_ada_rts
crate to be able to pass the Windows builds merge prerequisites of the
alire-index repo. However, it turns out the removed symbols are needed
to be able to compile HiRTOS.

Signed-off-by: J. German Rivera <jgrivera67@gmail.com>